Relog inside last room when disconnected
This is a behavior that really bothers me in the game: if you disconnect you can easily "cheat out" of any restraints you have on.

This fix should help mitigate the trick - the idea is that we keep track of the room you were in when you disconnect and we try to join it whenever you relog next. It will follow the normal room rules, so if it doesn't exist anymore or if it's full you will not be able to join it and the existing behavior is preserved. But if it's available, you will be back where you last were, making it harder to "cheat out" of restraints.

Disclaimer: I haven't tested or even run these changes, I mostly wanted to give the idea (although they seem simple enough that they should just work 😅)