Added tamper evidence system
Adds a tamper-evidence system which works in the following way:
-Accounts have a dictionary called TamperLock which stores information on which groups were last modified with a tamper-proof lock.
-When an account that has access to the player sends a tamperlock message, the server attaches the current time and name/membernumber of the person who sent the tamperlock message, and appends it to the current list
-If successful, it sends the tamper list to all players in the chatroom so that it updates on their screen
-both players must be in the same chatroom and the person applying must have permission on the target
-Values are synced on login and in character loading in chatrooms

There is one vulnerability, which is that someone can maliciously send TamperMessages to a player to reset the time on their tamper proof locks. However, the variable also stores the name and member number of whomever sent the update, so it will be clear to the owner (or whomever put the lock on there) that someone else tampered with the lock.