Fix a bug that resulted in a random lover disappearing
When a relationship is broken, and we were the target (ie. a self-lover)
both the other character's lovership update (with us as the target) and
our own lovership update would run, leading to a spurious removal of
another lover, as our calculated removal index wouldn't match what our
lovership after the "other update" ran.

Fix the bug by skipping our own update when we target ourselves.