Sync character on appearance change
The latest changes also added a line to delete AssetFamily from the
data of an account update, but Appearance wasn't in the list of
attributes that caused the character to be synced, so the server
stopped sending character sync messages in this case.

This removes AssetFamily from the list (since it can never be there
anyway now as we delete it) and replaces it with Appearance, which
should restore the server to sending the update it was sending
previously.